feat(css): checkbox: improve accessibility of disabled state
Use 'aria-disabled' instead of the 'disabled' attribute to make the component accessible to screen readers. The css selectors are made implementation-agnostic, only dependent on the '--disabled' classname modifier.

Known issue: The aria-disabled attribute does not by default disable "clickability" in most browsers today. Clicking with mouse is disabled here with css (pointer-events: none), but targeting and ticking the box with tab + space needs to be disabled with javascript (e.g. with preventDefault()).