feat(checkbox)!: fit label in focus outline
The custom checkbox is rewritten as a pseudo-element as part of the label to include the label (the total clickable area) inside the focus outline visible when using keyboard navigation. This involves a change in the component HTML composition, most notably using a separate label, istead of wrapping the input in the label.