refactor: separate HTML templating and component JS
Split the "packages/components" directory containing HTML templating with Storybook stories and component level scripts in to two directories "packages/html" and "packages/js".

"packages/html": Contains HTML templating with accompanying core component stories. This is the reference markup implementation using the CSS from "packages/css" and JS from "packages/js". The directory is not planned as an NPM-package, but is kept in the "packages/" directory for practical reasons, if any CI/CD/build-scripting is needed, and thus could be included in a "package.json" file and will be automatically run in the lerna building process.

"packages/js": Contains JS needed for some core components. Planned as NPM-package. Configuration files are adapted to the new project plan (package.json / LICENSE / readme.md).